Refactoring: Move profiler.h and profiler.cpp to :lib-core and solve a LOT of gradle build errors.
