+++ lib-core/build.gradle
@@ -11,6 +11,33 @@ android {
         versionName "1.0"
 
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+
+        ndk {
+            // "x86", "x86_64", "armeabi", "armeabi-v7a", "arm64-v8a"
+            abiFilters "armeabi-v7a"
+        }
+        externalNativeBuild {
+            cmake {
+                arguments "-DANDROID_PLATFORM=android-16",
+                          "-DANDROID_TOOLCHAIN=clang",
+                          "-DANDROID_STL=c++_shared",
+                          "-DANDROID_CPP_FEATURES=rtti exceptions"
+                // Enables exception-handling support.
+//                cppFlags "-fexceptions -std=c++11"
+            }
+        }
+    }
+
+    // Let gradle pack the shared library into apk.
+    sourceSets {
+        main {
+            jniLibs.srcDirs = ['src/main/cppLibs/lib']
+        }
+    }
+    externalNativeBuild {
+        cmake {
+            path 'src/main/cpp/CMakeLists.txt'
+        }
     }
 
     buildTypes {
