+++ lib-core/src/main/cpp/profiler.cpp
@@ -1,4 +1,4 @@
-#include "profiler.h"
+#include <my/profiler.h>
 #include <android/log.h>
 
 #define LOGI(...) \
