+++ lib-dlib/src/main/cpp/CMakeLists.txt
@@ -1,8 +1,13 @@
 cmake_minimum_required(VERSION 3.4.1)
 
-# Configure import libs
-set(lib_dir ${CMAKE_SOURCE_DIR}/../cppLibs)
-message(STATUS "lib_dir=${lib_dir}")
+# Configure build library name.
+set(TARGET_NAME dlib_jni)
+
+# Configure install folder.
+set(INSTALL_DIR ${CMAKE_SOURCE_DIR}/../../../../lib-distribution/)
+
+# Configure import libs.
+set(LIB_DIR ${CMAKE_SOURCE_DIR}/../cppLibs)
 
 # Link to libdlib.so.
 # Shared lib will also be tucked into APK and sent to target
@@ -11,29 +16,71 @@ message(STATUS "lib_dir=${lib_dir}")
 # the most valuable thing of this sample, the rest are pretty much normal cmake
 add_library(dlib SHARED IMPORTED)
 set_target_properties(dlib PROPERTIES IMPORTED_LOCATION
-    ${lib_dir}/dlib/lib/${ANDROID_ABI}/libdlib.so)
+    ${LIB_DIR}/dlib/lib/${ANDROID_ABI}/libdlib.so)
 
 # Link to libprotobuf-lite.so.
 add_library(protobuf SHARED IMPORTED)
 set_target_properties(protobuf PROPERTIES IMPORTED_LOCATION
     ${CMAKE_SOURCE_DIR}/../../../../lib-protobuf/cppLibs/lib/${ANDROID_ABI}/libprotobuf-lite-3.2.0.so)
 
+# Link to libmy_core_jni.so.
+add_library(my_core_jni SHARED IMPORTED)
+set_target_properties(my_core_jni PROPERTIES IMPORTED_LOCATION
+    ${INSTALL_DIR}/my_core_jni/lib/${ANDROID_ABI}/libmy_core_jni.so)
+
 # Build application's shared lib
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -fexceptions -std=c++11")
 
-add_library(dlib_jni SHARED
+add_library(${TARGET_NAME} SHARED
             dlib-jni.cpp
-            data/messages.pb.cc
-            util/profiler.cpp)
+            data/messages.pb.cc)
 
-target_include_directories(dlib_jni PRIVATE
-                           ${CMAKE_SOURCE_DIR}
+target_include_directories(${TARGET_NAME} PRIVATE
+                           ${INSTALL_DIR}/my_core_jni/include
                            ${CMAKE_SOURCE_DIR}/../../../../lib-protobuf/cppLibs/include
-                           ${lib_dir}/dlib/include)
+                           ${CMAKE_SOURCE_DIR}
+                           ${LIB_DIR}/dlib/include)
 
-target_link_libraries(dlib_jni
+target_link_libraries(${TARGET_NAME}
                       dlib
+                      my_core_jni
                       protobuf
                       android
                       jnigraphics
                       log)
+
+# Make install directory.
+file(MAKE_DIRECTORY ${INSTALL_DIR}/${TARGET_NAME})
+file(MAKE_DIRECTORY ${INSTALL_DIR}/${TARGET_NAME}/include)
+file(MAKE_DIRECTORY ${INSTALL_DIR}/${TARGET_NAME}/lib)
+
+# Copy out lib header file...
+add_custom_command(TARGET ${TARGET_NAME} POST_BUILD
+
+                   COMMAND "${CMAKE_COMMAND}" -E
+                   copy_directory "${LIB_DIR}/dlib/include"
+                   "${INSTALL_DIR}/${TARGET_NAME}/include/"
+
+                   COMMENT "Copying ${TARGET_NAME} headers to output directory")
+
+# Copy out the lib binary.
+# FIXME: ExternalNativeBuild breaks on copying STL. It might be fixed in the
+# FIXME: next Android Studio release (2.5.x).
+# FIXME: Issue thread: https://github.com/android-ndk/ndk/issues/334
+add_custom_command(TARGET ${TARGET_NAME} POST_BUILD
+
+                   # Copy cppLibs/dlib/lib/*.so
+                   COMMAND "${CMAKE_COMMAND}" -E
+                   copy_directory "${LIB_DIR}/dlib/lib/${ANDROID_ABI}"
+                   "${INSTALL_DIR}/${TARGET_NAME}/lib/${ANDROID_ABI}"
+
+                   # Copy generated *.so.
+                   COMMAND "${CMAKE_COMMAND}" -E
+                   copy_directory "${CMAKE_SOURCE_DIR}/../../../build/intermediates/cmake/${CMAKE_BUILD_TYPE}/obj/${ANDROID_ABI}"
+                   "${INSTALL_DIR}/${TARGET_NAME}/lib/${ANDROID_ABI}"
+
+                   COMMENT "Copying ${TARGET_NAME} shared library to output directory")
+# set_target_properties(${TARGET_NAME}
+#                       PROPERTIES
+#                       LIBRARY_OUTPUT_DIRECTORY
+#                       "${INSTALL_DIR}/${TARGET_NAME}/lib/${ANDROID_ABI}")
