+++ lib-dlib/src/main/java/com/my/jni/dlib/FaceLandmarksDetector.java
@@ -46,6 +46,16 @@ public class FaceLandmarksDetector {
 
         // TODO: Load library in worker thread?
         try {
+            System.loadLibrary("my_core_jni");
+            Log.d("jni", "my_core_jni.so is loaded");
+        } catch (UnsatisfiedLinkError error) {
+            throw new RuntimeException(
+                "\"my_core_jni\" not found; check that the correct native " +
+                "libraries are present in the APK.");
+        }
+
+        // TODO: Load library in worker thread?
+        try {
             System.loadLibrary("protobuf-lite-3.2.0");
             Log.d("jni", "libprotobuf-lite-3.2.0.so is loaded");
         } catch (UnsatisfiedLinkError error) {
